Image for processing works in Internet Explorer (before did work in Firefox only)

+++ OpenXava/src/org/openxava/calculators/NextLongCalculator.java
@@ -32,7 +32,7 @@ public class NextLongCalculator implements IJDBCCalculator {
 		if (XSystem.onClient()) {
 			Object r = Server.calculate(this, getPackageName());			
 			if (r instanceof Number) {
-				return ((Number) r).intValue();
+				return ((Number) r).longValue();
 			}
 			return 0;
 		}
